+++ lib/src/abstraction_layer/rw/model_writer.dart
@@ -13,6 +13,7 @@ class ModelWriter {
 
 class CanvasModelWriter extends ModelWriter
     with ComponentWriter, LinkWriter, ConnectionWriter {
+  /// Allows you to change the model.
   CanvasModelWriter(CanvasModel canvasModel, CanvasState canvasState)
       : super(canvasModel, canvasState);